+++ Breakout/Breakout/Graphics/GraphicsManager.cpp
@@ -105,7 +105,7 @@ ICamera* GraphicsManager::GetICamera()
     return m_ICamera;
 }
 
-void GraphicsManager::AddTextObject(std::string* _text,float* _scale, unsigned int* _color,int* _x,int* _y,int _id)
+void GraphicsManager::AddTextObject(std::string* _text,float* _scale, unsigned int* _color,float* _x,float* _y,int _id)
 {
     m_IGraphics->AddTextObject(_text,_scale,_color,_x,_y,_id);
 }
