+++ Breakout/Breakout/Graphics/GraphicsManager.h
@@ -59,7 +59,7 @@ public:
 	void Add2DTexture(int _id, std::string _path, float *_x, float *_y, float *_width, float *_height);
 	void Remove2DTexture(int _id);
         void RemovePointLight(int _id);
-        void AddTextObject(std::string* _text,float* _scale, unsigned int* _color,int* _x,int* _y,int _id);
+        void AddTextObject(std::string* _text,float* _scale, unsigned int* _color,float* _x,float* _y,int _id);
         void RemoveTextObject(int _id);
